refacto(pix-table)!: Simplify pix-table usage
