--- addon/components/pix-table-basic-column.hbs
@@ -1,8 +1,0 @@
-<PixTableColumn @context={{@context}} class={{this.typeClass}}>
-  <:header>
-    {{@name}}
-  </:header>
-  <:cell>
-    {{@value}}
-  </:cell>
-</PixTableColumn>
