+++ addon/components/pix-table-column.hbs
@@ -3,7 +3,7 @@
     {{yield to="header"}}
   </th>
 {{else}}
-  <td ...attributes>
+  <td ...attributes class={{this.typeClass}}>
     {{yield to="cell"}}
   </td>
 {{/if}}
