+++ addon/styles/_pix-table-column.scss
@@ -1,4 +1,4 @@
-td.pix-table-basic-column {
+.pix-table-column {
   &--number{
     text-align: right;
   }
