+++ addon/styles/_pix-table.scss
@@ -1,4 +1,4 @@
-@import 'pix-table-basic-column';
+@import 'pix-table-column';
 
 .pix-table {
   width: 100%;
