--- app/components/pix-table-basic-column.js
@@ -1,1 +1,0 @@
-export { default } from '@1024pix/pix-ui/components/pix-table-basic-column';
