--- app/stories/pix-table-basic-column.mdx
@@ -1,36 +1,0 @@
-import { Meta, Story, ArgTypes } from '@storybook/blocks';
-
-import * as ComponentStories from './pix-table-basic-column.stories';
-
-<Meta of={ComponentStories} />
-
-# PixTableBasicColumn
-
-Un composant enfant du PixTable qui affiche une colonne de type `text` ou `number`
-
-<Story of={ComponentStories.Default} height={400} />
-
-## Usage
-
-```html
-<PixTable @data={{this.data}} @caption={{this.caption}}>
-  <:columns as |row context|>
-    <PixTableBasicColumn
-      @context={{context}}
-      @type='text'
-      @name='Nom'
-      @value={{row.name}}
-    />
-    <PixTableBasicColumn
-      @context={{context}}
-      @type='number'
-      @name='Âge'
-      @value={{row.age}}
-    />
-  </:columns>
-</PixTable>
-```
-
-## Arguments
-
-<ArgTypes />
